+++ packages/expo-updates/android/src/main/java/expo/modules/updates/statemachine/UpdatesStateContext.kt
@@ -1,5 +1,6 @@
 package expo.modules.updates.statemachine
 
+import android.os.Bundle
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.WritableMap
 import org.json.JSONObject
@@ -50,31 +51,41 @@ data class UpdatesStateContext(
    */
   val writableMap: WritableMap
     get() {
-      val contextMap = Arguments.createMap()
-      contextMap.putBoolean("isUpdateAvailable", isUpdateAvailable)
-      contextMap.putBoolean("isUpdatePending", isUpdatePending)
-      contextMap.putBoolean("isRollback", isRollback)
-      contextMap.putBoolean("isChecking", isChecking)
-      contextMap.putBoolean("isDownloading", isDownloading)
-      contextMap.putBoolean("isRestarting", isRestarting)
-      if (latestManifest != null) {
-        contextMap.putString("latestManifestString", latestManifest.toString())
-      }
-      if (downloadedManifest != null) {
-        contextMap.putString("downloadedManifestString", downloadedManifest.toString())
-      }
-      if (checkError != null) {
-        val errorMap = Arguments.createMap()
-        errorMap.putString("message", checkError.message)
-        contextMap.putMap("checkError", errorMap)
-      }
-      if (downloadError != null) {
-        val errorMap = Arguments.createMap()
-        errorMap.putString("message", downloadError.message)
-        contextMap.putMap("downloadError", errorMap)
-      }
       val result = Arguments.createMap()
-      result.putMap("context", contextMap)
+      result.putMap("context", Arguments.fromBundle(bundle))
       return result
+    }
+
+  /**
+   * Creates a Bundle to be returned to JS on a call to nativeStateMachineContext()
+   */
+  val bundle: Bundle
+    get() {
+      return Bundle().apply {
+        putBoolean("isUpdateAvailable", isUpdateAvailable)
+        putBoolean("isUpdatePending", isUpdatePending)
+        putBoolean("isRollback", isRollback)
+        putBoolean("isChecking", isChecking)
+        putBoolean("isDownloading", isDownloading)
+        putBoolean("isRestarting", isRestarting)
+        if (latestManifest != null) {
+          putString("latestManifestString", latestManifest.toString())
+        }
+        if (downloadedManifest != null) {
+          putString("downloadedManifestString", downloadedManifest.toString())
+        }
+        if (checkError != null) {
+          val errorMap = Bundle().apply {
+            putString("message", checkError.message)
+          }
+          this.putBundle("checkError", errorMap)
+        }
+        if (downloadError != null) {
+          val errorMap = Bundle().apply {
+            putString("message", downloadError.message)
+          }
+          putBundle("downloadError", errorMap)
+        }
+      }
     }
 }